+++ vehicle/vw/weconnect/provider.go
@@ -1,4 +1,4 @@
-package id
+package weconnect
 
 import (
 	"errors"
