+++ vehicle/vw/weconnect/types.go
@@ -1,4 +1,4 @@
-package id
+package weconnect
 
 import (
 	"errors"
